+++ RunSmart/app/src/main/java/com/project_wombat/runsmart/RunData.java
@@ -11,10 +11,10 @@ public class RunData {
 	private double speed;
 	private double longitude;
 	private double latitude;
-	private int time_elapsed;
+	private long time_elapsed;
 
 	public RunData(){}
-	public RunData(Date time_stamp, double speed, double longitude, double latitude, int time_elapsed)
+	public RunData(Date time_stamp, double speed, double longitude, double latitude, long time_elapsed)
 	{
 		this.time_stamp = time_stamp.getTime()/1000;
 		this.speed = speed;
@@ -27,11 +27,11 @@ public class RunData {
 	public void setSpeed(double speed) { this.speed = speed; }
 	public void setLongitude(double longitude) { this.longitude = longitude; }
 	public void setLatitude(double latitude) { this.latitude = latitude; }
-	public void setTimeElapsed(int time_elapsed) { this.time_elapsed = time_elapsed; }
+	public void setTimeElapsed(long time_elapsed) { this.time_elapsed = time_elapsed; }
 
 	public long getTimeStamp() { return time_stamp; }
 	public double getSpeed() { return speed; }
 	public double getLongitude() { return longitude; }
 	public double getLatitude() { return latitude; }
-	public int getTimeElapsed() { return time_elapsed; }
+	public long getTimeElapsed() { return time_elapsed; }
 }
